Etablerer pakke web for web-laget
