+++ src/main/java/no/bankid/oidc/web/LoginServlet.java
@@ -1,6 +1,8 @@
-package no.bankid.oidc; /**
+package no.bankid.oidc.web; /**
  * Created by kristofferskaret on 20.01.2017.
  */
+
+import no.bankid.oidc.BankIdOIDCClient;
 
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -8,19 +10,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = {"/callback"}, loadOnStartup = 1)
-public class CallBackServlet extends HttpServlet {
+@WebServlet(urlPatterns = {"/login"}, loadOnStartup = 1)
+public class LoginServlet extends HttpServlet {
 
     @Override
     public void doGet(HttpServletRequest request, HttpServletResponse response)
             throws IOException {
 
-        String code = request.getParameter("code");
+        String redirectUrl = BankIdOIDCClient.getInstance().createAuthenticationUrl();
 
-        User user = BankIdOIDCClient.getInstance().endAuthentication(code);
-
-        request.getSession().setAttribute("user", user);
-
-        response.sendRedirect("/");
+        response.sendRedirect(redirectUrl);
     }
 }
